Fix slide progression
Youtube slides was failing to progress to next slide on video
ended event.

It was also failing when two youtube slides was placed afte each
other. Wrapping in setimeout with fadetime like elsewhere seems
to fix this.

Without slide.player.destroy() it would only play on initial load
and just show ending in subsequent playbakcs of the slide.